# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next Olshop API",
    "description": "REST API for Next Olshop — Single Store Online Shop untuk UMKM. Digunakan oleh Flutter mobile app.\n\n## Authentication\nSemua endpoint yang membutuhkan auth menggunakan Bearer token dari `/api/auth/login`.\n\n## Response Format\n- **Success**: `{ success: true, data: T }`\n- **Error**: `{ success: false, error: string }`\n- **Pagination**: `{ success: true, data: T[], meta: { page, limit, total, totalPages } }`\n\n## Catatan\n- Harga dalam Rupiah (IDR), disimpan sebagai string decimal (`\"299000.00\"`)\n- Tanggal dalam format ISO 8601 (UTC)\n- Berat produk dalam gram",
    "version": "1.0.0",
    "contact": {
      "name": "Next Olshop",
      "url": "https://github.com/codewithbahri"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token dari POST /api/auth/login"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 2 },
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "format": "email", "example": "john@example.com" },
          "phone": { "type": "string", "nullable": true, "example": "08123456789" },
          "role": { "type": "string", "enum": ["customer", "admin"], "example": "customer" }
        }
      },
...
      "Address": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 2 },
          "recipientName": { "type": "string", "example": "John Doe" },
          "phone": { "type": "string", "example": "08123456789" },
          "address": { "type": "string", "example": "Jl. Raya Sudirman No. 123" },
          "detail": { "type": "string", "nullable": true, "example": "Apt 456, Gedung B" },
          "province": { "type": "string", "example": "DKI Jakarta" },
          "provinceId": { "type": "string", "nullable": true, "example": "31" },
          "city": { "type": "string", "example": "Jakarta Pusat" },
          "cityId": { "type": "string", "nullable": true, "example": "3171" },
          "district": { "type": "string", "example": "Menteng" },
          "districtId": { "type": "string", "nullable": true, "example": "317171" },
          "postalCode": { "type": "string", "example": "10310" },
          "latitude": { "type": "string", "nullable": true, "example": "-6.2088" },
          "longitude": { "type": "string", "nullable": true, "example": "106.8456" },
          "label": { "type": "string", "nullable": true, "example": "Rumah" },
          "isDefault": { "type": "boolean", "example": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Error message" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Validasi gagal" },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            },
            "example": { "email": ["Email tidak valid"], "password": ["Password minimal 6 karakter"] }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Login dengan email & password, return JWT token + user data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "john@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIs..." },
                        "user": { "$ref": "#/components/schemas/User" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "401": {
            "description": "Email atau password salah",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/addresses": {
      "get": {
        "tags": ["Addresses"],
        "summary": "List Addresses",
        "description": "Ambil semua alamat pengiriman user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List alamat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Address" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "post": {
        "tags": ["Addresses"],
        "summary": "Create Address",
        "description": "Tambah alamat pengiriman baru",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["recipientName", "phone", "address", "province", "city", "district", "postalCode"],
                "properties": {
                  "recipientName": { "type": "string", "example": "John Doe" },
                  "phone": { "type": "string", "example": "08123456789" },
                  "address": { "type": "string", "example": "Jl. Raya Sudirman No. 123" },
                  "detail": { "type": "string", "example": "Apt 456, Gedung B" },
                  "province": { "type": "string", "example": "DKI Jakarta" },
                  "provinceId": { "type": "string", "example": "31" },
                  "city": { "type": "string", "example": "Jakarta Pusat" },
                  "cityId": { "type": "string", "example": "3171" },
                  "district": { "type": "string", "example": "Menteng" },
                  "districtId": { "type": "string", "example": "317171" },
                  "postalCode": { "type": "string", "example": "10310" },
                  "latitude": { "type": "string", "example": "-6.2088" },
                  "longitude": { "type": "string", "example": "106.8456" },
                  "label": { "type": "string", "example": "Rumah" },
                  "isDefault": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alamat dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Address" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/addresses/{id}": {
      "get": {
        "tags": ["Addresses"],
        "summary": "Get Address",
        "description": "Ambil detail alamat berdasarkan ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Address ID" }
        ],
        "responses": {
          "200": {
            "description": "Detail alamat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Address" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Alamat tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "put": {
        "tags": ["Addresses"],
        "summary": "Update Address",
        "description": "Update alamat pengiriman",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Address ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipientName": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "detail": { "type": "string" },
                  "province": { "type": "string" },
                  "provinceId": { "type": "string" },
                  "city": { "type": "string" },
                  "cityId": { "type": "string" },
                  "district": { "type": "string" },
                  "districtId": { "type": "string" },
                  "postalCode": { "type": "string" },
                  "latitude": { "type": "string" },
                  "longitude": { "type": "string" },
                  "label": { "type": "string" },
                  "isDefault": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alamat diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Address" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Alamat tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["Addresses"],
        "summary": "Delete Address",
        "description": "Hapus alamat pengiriman",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Address ID" }
        ],
        "responses": {
          "200": {
            "description": "Alamat dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Alamat tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/addresses/{id}/default": {
      "put": {
        "tags": ["Addresses"],
        "summary": "Set Default Address",
        "description": "Set alamat sebagai alamat utama (default). Alamat lain akan di-unset.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Address ID" }
        ],
        "responses": {
          "200": {
            "description": "Alamat utama diset",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Alamat tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}